requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-nested-routers==0.93.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==3.20.1
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
Werkzeug==2.2.2
